requirements.txt
streamlit
openai
gspread
oauth2client
langdetect
google_trans_new
python-dotenv
#